collision voor items zijn ingevoegt

+++ Pigit/Pigit/Movement/NPCMoveCommands/MoveCommandWalkNPC.cs
@@ -21,7 +21,7 @@ namespace Pigit.Movement.NPCMoveCommands
         private double timer;
         private bool isSetTimer = false;
 
-        public MoveCommandWalkNPC(IPlayerObject player, Level level,int jumpHeight = 4,int walkspeed= 2) : base(player, level, jumpHeight, walkspeed)
+        public MoveCommandWalkNPC(IPlayerObject player, Level level,float jumpHeight = 4,float walkspeed= 2) : base(player, level, jumpHeight, walkspeed)
         {
 
         }
